+++ src/main/java/me/kbh/elktoy/doc/presentation/SampleController.java
@@ -1,4 +1,4 @@
-package me.kbh.elktoy.member.presentation;
+package me.kbh.elktoy.doc.presentation;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
